change fabs to fabsf

+++ test_gpt2.c
@@ -7,7 +7,7 @@ int check_tensor(float *a, float *b, int n, char* label) {
     int ok = 1;
     printf("%s\n", label);
     for (int i = 0; i < n; i++) {
-        if (fabs(a[i] - b[i]) <= 1e-2) {
+        if (fabsf(a[i] - b[i]) <= 1e-2) {
             if (i < print_upto) { printf("OK "); }
         } else {
             if (i < print_upto) { printf("NOT OK "); }
@@ -91,7 +91,7 @@ int main(int argc, char *argv[]) {
                 if(i < 3) {
                     printf("%f %f\n", expected_logits[i], model.acts.logits[i]);
                 }
-                if (fabs(expected_logits[i] - model.acts.logits[i]) >= 1e-2) {
+                if (fabsf(expected_logits[i] - model.acts.logits[i]) >= 1e-2) {
                     printf("MISMATCH AT INDEX %d: ", i);
                     printf("%f %f\n", expected_logits[i],model.acts.logits[i]);
                     logits_ok = 0;
@@ -103,7 +103,7 @@ int main(int argc, char *argv[]) {
             allok = allok && logits_ok;
 
             // compare the achieved loss
-            if (fabs(model.mean_loss - *expected_loss) >= 1e-2) {
+            if (fabsf(model.mean_loss - *expected_loss) >= 1e-2) {
                 printf("LOSS MISMATCH: %f %f\n", model.mean_loss, *expected_loss);
                 allok = 0;
             } else {
@@ -156,7 +156,7 @@ int main(int argc, char *argv[]) {
     };
     // compare
     for (int i = 0; i < 10; i++) {
-        if (fabs(losses[i] - expected_losses[i]) >= 1e-2) {
+        if (fabsf(losses[i] - expected_losses[i]) >= 1e-2) {
             printf("LOSS MISMATCH AT STEP %d: %f %f\n", i, losses[i], expected_losses[i]);
             allok = 0;
         } else {
